lab7: Done lab7_3

+++ lab7/lab7_3testContainers/src/main/java/pt/ua/deti/lab7_3testContainers/repository/BookRepo.java
@@ -5,7 +5,7 @@ import pt.ua.deti.lab7_3testContainers.model.Book;
 
 import java.util.List;
 
-public interface BookRepo extends JpaRepository<Book, Integer>{
+public interface BookRepo extends JpaRepository<Book, Long>{
     List<Book> findAll();
     List<Book> findByTitle(String title);
     List<Book> findByAuthor(String author);
